requirements.txt
absl-py==2.1.0
accelerate==1.1.1
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
altair==5.5.0
ansible==9.5.1
ansible-core==2.16.6
ansible-runner==2.4.0
ansible_rulebook==1.1.0
appnope==0.1.4
asttokens==3.0.0
astunparse==1.6.3
attrs==23.2.0
autopep8==2.3.1
bcrypt==4.1.2
blinker==1.7.0
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.0
comm==0.2.2
contourpy==1.3.1
cryptography==42.0.5
cycler==0.12.1
databricks-sdk==0.37.0
datasets==3.1.0
debugpy==1.8.12
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.7
docker==7.1.0
docutils==0.21.2
dpath==2.1.6
drools_jpy==0.3.9
evaluate==0.4.3
executing==2.2.0
fairlearn==0.11.0
Farama-Notifications==0.0.4
filelock==3.13.1
flake8==7.1.1
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-HTTPAuth==4.8.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
flatbuffers==24.3.25
fonttools==4.55.0
frozenlist==1.4.1
fsspec==2024.9.0
gast==0.6.0
gcsfs==2024.9.0.post1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.18.0
google-auth==2.29.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-pubsub==2.21.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.0
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==23.0.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.0.0
h5py==3.12.1
huggingface-hub==0.26.2
idna==3.6
importlib_metadata==8.5.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.1.2
janus==1.0.0
jedi==0.19.2
Jinja2==3.1.3
joblib==1.4.2
jpy==0.16.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.7.0
kiwisolver==1.4.7
libclang==18.1.1
lockfile==0.12.2
Mako==1.3.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
matrix-mdp-gym==1.1.1
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.4.1
mlflow==2.17.2
mlflow-skinny==2.17.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mysql-connector-python==8.2.0
namex==0.0.8
narwhals==1.15.1
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.0.2
oauthlib==3.2.2
opentelemetry-api==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
opt_einsum==3.4.0
optree==0.13.1
packaging==23.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.0.0
pluggy==1.4.0
prompt_toolkit==3.0.50
proto-plus==1.23.0
protobuf==4.21.12
psutil==6.1.0
psycopg2==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.1
pycparser==2.21
pydeck==0.9.1
pyflakes==3.2.0
pygame==2.6.1
Pygments==2.18.0
PyMySQL==1.1.0
pyparsing==3.1.2
pytest==8.0.2
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.2
PyYAML==6.0.1
pyzmq==26.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
resolvelib==1.0.1
rich==13.9.4
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.5
scapy==2.6.1
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.27
sqlparse==0.5.2
stack-data==0.6.3
streamlit==1.40.2
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
tokenizers==0.20.3
toml==0.10.2
torch==2.5.1
tornado==6.4.2
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.2
typing_extensions==4.10.0
tzdata==2024.2
urllib3==2.2.1
virtualenv==20.24.7
watchdog==4.0.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.21.0
